Always use 32 bit CRC
* src/cpiohdr.h (cpio_file_stat) <c_chksum>: Change type to
uint32_t.
* src/extern.h (crc): Change type to uint32_t.
* src/global.c: Likewise.
* src/copyout.c (read_for_checksum): Return uint32_t.
* src/copyin.c: Fix printf formats.

+++ src/cpiohdr.h
@@ -123,7 +123,7 @@ struct cpio_file_stat /* Internal representation of a CPIO header */
   long c_rdev_maj;
   long c_rdev_min;
   size_t c_namesize;
-  unsigned long c_chksum;
+  uint32_t c_chksum;
   char *c_name;
   char *c_tar_linkname;
 };
