+++ src/global.c
@@ -144,7 +144,7 @@ char *archive_name = NULL;
 char *rsh_command_option = NULL;
 
 /* CRC checksum.  */
-unsigned long crc;
+uint32_t crc;
 
 /* Input and output buffers.  */
 char *input_buffer, *output_buffer;
